今天工作重点是在完善显示界面上面，搞定了对话框的阴影效果以及后背景颜色透明度，基本上就这些，还有多功能复选的checkbox已经差不多了，后续求优化

+++ Msafe/MSafe/src/org/com/lix_/util/UiUtils.java
@@ -30,7 +30,7 @@ public class UiUtils {
 		return (int) (pxValue / scale + 0.5f);
 	}
 
-	static public int dipToPx(Context context, int dip) {
+	static public int dipToPx(Context context, float dip) {
 		return (int) (dip * getScreenDensity(context) + 0.5f);
 	}
